chore: move reset password module to worker
