+++ packages/apps/human-app/frontend/src/modules/worker/reset-password/index.ts
@@ -1,0 +1,2 @@
+export * from './reset-password-success.page';
+export * from './reset-password.page';
